==> src/1.1-Feature-engineering-close-buildings.ipynb <==
%load_ext autoreload
%autoreload 2
%cd C:\MAD4AG
%matplotlib inline

import pandas as pd
import numpy as np
from tqdm.auto import tqdm
from sklearn.cluster import DBSCAN
import geopandas
import folium
from folium.plugins import HeatMapWithTime
import warnings

warnings.filterwarnings('ignore')


file_name = f'./dbs/big_geodata_se/stops.parquet'

#Read each batchs file
df = pd.read_parquet(file_name)

# DBSCAN to detect jumps


def add_individual_clusters(data):
    # represent points consistently as (lat, lon)
    coords = data[['lat', 'lng']].values

    # define the number of kilometers in one radian
    kms_per_radian = 6371.0088

    # define epsilon as 500 kilometers, converted to radians for use by haversine
    epsilon = 200 / kms_per_radian

    db = DBSCAN(eps=epsilon, min_samples=2, algorithm='ball_tree', metric='haversine').fit(np.radians(coords))
    cluster_labels = db.labels_
    data['cluster'] = cluster_labels + 1
    return data




tqdm.pandas()
df = df.groupby('uid').progress_apply(add_individual_clusters)

print('The number of people: ', len(df['uid'].unique()))
print('The number of people having jump: ', len(df['uid'][df['cluster']>1].unique()))

df['uid'] = df['uid']+'-'+df['cluster'].astype(str)

df.drop(columns=['cluster'], inplace=True)

# remove holidays and weekend stops

df = df[df.holiday_s != 1]
df = df[df.weekday_s == 1]

# DBSCAN to detect clusters

def add_individual_clusters(data):
    # represent points consistently as (lat, lon)
    coords = data[['lat', 'lng']].values

    # define the number of kilometers in one radian
    kms_per_radian = 6371.0088

    # define epsilon as 0.1 kilometers, converted to radians for use by haversine
    epsilon = 0.1 / kms_per_radian

    db = DBSCAN(eps=epsilon, min_samples=1, algorithm='ball_tree', metric='haversine').fit(np.radians(coords))
    cluster_labels = db.labels_
    data['cluster'] = cluster_labels + 1
    return data


tqdm.pandas()
df = df.groupby('uid').progress_apply(add_individual_clusters)

# the clusters in one individual's different jumps

df[df['uid'] == '9b0b36c9-b1c0-444e-8708-e2751915a587-1']

df[df['uid'] == '9b0b36c9-b1c0-444e-8708-e2751915a587-2']

## Drop individuals with only one cluster

cluster_count = df.groupby('uid')['cluster'].nunique().reset_index(name='cluster_count')

ppl_with_more_cluster = cluster_count['uid'][cluster_count.cluster_count>1].tolist()

df = df[df['uid'].isin(ppl_with_more_cluster)]

## Intersect MAD data and DeSO data

gdf = geopandas.GeoDataFrame(df, geometry=geopandas.points_from_xy(df.lng, df.lat), crs="EPSG:4326")

DeSO = geopandas.read_file(f'C:/Synthetic_population_new/caglar/synthetic_sweden/input/deso_statistik_shp/Bef_Kon_region.shp')

print(DeSO.crs)

DeSO.to_crs(4326, inplace=True)
print(DeSO.crs)

gdf = gdf.sjoin(DeSO[['Deso','geometry']], how="left")

gdf.drop("index_right", axis=1, inplace=True)
#out of sweden
print('Stops outside Sweden: ', len(gdf[gdf['Deso'].isna()]))

gdf = gdf.dropna(subset=['Deso'])

df = pd.DataFrame(gdf.drop(columns='geometry'))


## Drop individuals with only one cluster

cluster_count = df.groupby('uid')['cluster'].nunique().reset_index(name='cluster_count')

ppl_with_more_cluster = cluster_count['uid'][cluster_count.cluster_count>1].tolist()

df = df[df['uid'].isin(ppl_with_more_cluster)]

## Calculate mean of clusters and snap to the closest building


df_cluster = df.groupby(['uid','cluster'])[['lat','lng']].mean().reset_index()


## Snap the cluster centre to the closest building.

# building_point = geopandas.read_file(f'C:/Synthetic_population_new/caglar/synthetic_sweden/input/Geodatabase.gdb', layer= 'Buildings_SpatJ_SU_1km_DESO_poi_wgs84')
#
# building_point = pd.DataFrame(building_point[['TARGET_FID','ANDAMAL_1',
#        'Deso', 'KNKOD',  'POINT_X_sweref99',
#        'POINT_Y_sweref99', 'POINT_X', 'POINT_Y']])
# building_point = building_point[building_point.ANDAMAL_1 != 699]
# building_point.rename(columns={'POINT_Y': 'lat', 'POINT_X': 'lng'}, inplace=True)

#HERE
# df_cluster.to_parquet(f'./dbs/intermediate/df_cluster.parquet')
# building_point.to_parquet(f'./dbs/intermediate/building_point.parquet')


# df_cluster = pd.read_parquet(f'./dbs/intermediate/df_cluster.parquet')
building_point = pd.read_parquet(f'./dbs/intermediate/building_point.parquet')


from lib.nearest_point import nearest_neighbor

closest_build = nearest_neighbor(df_cluster, building_point, return_dist=True)


closest_build= closest_build[['TARGET_FID', 'ANDAMAL_1', 'Deso', 'KNKOD', 'POINT_X_sweref99','POINT_Y_sweref99', 'lat', 'lng', 'distance']]

closest_build.rename(columns={'lat':'building_lat', 'lng':'building_lng'}, inplace=True)

df_cluster = df_cluster.join(closest_build)


df_cluster= df_cluster[['uid', 'cluster', 'lat', 'lng', 'TARGET_FID',
       'ANDAMAL_1', 'Deso', 'KNKOD', 'building_lng', 'building_lat', 'distance']]

df_cluster.rename(columns={'lat':'cluster_lat', 'lng':'cluster_lng'}, inplace=True)



### If the distance between the closest building and the cluster centre is more than 200m, use the coordinates of the cluster.

#HERE

df_cluster.cluster_lat = np.where(df_cluster['distance'] <= 0.2 , df_cluster.building_lat, df_cluster.cluster_lat)

df_cluster.cluster_lng = np.where(df_cluster['distance'] <= 0.2 , df_cluster.building_lng, df_cluster.cluster_lng)

df_cluster.ANDAMAL_1 = np.where(df_cluster['distance'] >= 0.2 , 0, df_cluster.ANDAMAL_1)

df_cluster.ANDAMAL_1.value_counts(normalize=True)


df = pd.merge(df, df_cluster[['uid', 'cluster', 'cluster_lat', 'cluster_lng', 'building_lat','building_lng', 'ANDAMAL_1', 'distance']], on=['uid', 'cluster'], how='left' )

#HERE
df.to_parquet(f'./dbs/intermediate/stops_1_new.parquet')
