==> src/1.3-Descriptive-analysis.ipynb <==
%cd C:\MAD4AG
%matplotlib inline

import pandas as pd
import numpy as np

import warnings

warnings.filterwarnings('ignore')

file_name = f'./dbs/intermediate/stops_1_new.parquet'

#Read each batchs file
df = pd.read_parquet(file_name)



df['id'] = df['uid'].str[:-2]

print('The number of people: ', len(df['id'].unique()))
print('The number of activity clusters: ', len(df['uid'].unique()))
print('The number of stops in weekday and non-holiday: ', len(df))

aaa = df.head(100)

bbb = df[df['id']=='d3ad9a03-e1a7-4229-8967-f187c3185c92']

df['m-d']= df['localtime'].dt.strftime('%m-%d')
active_days = df.groupby(['uid'])['m-d'].nunique().reset_index(name='active_days')
print('No. of active weekdays in activity clusters')
active_days['active_days'].quantile([.0, 0.25, .5, 0.75, 1])

median_dur = df.groupby(['uid'])['dur'].median().reset_index(name='median_dur')
print('Median duration (min)')
median_dur['median_dur'].quantile([.0, 0.25, .5, 0.75, 1])

loc_count = df.groupby(['uid']).size().reset_index(name='loc_count')
print('No. of stop at individuals\' activity clusters')
loc_count['loc_count'].quantile([.0, 0.25, .5, 0.75, 1])

uniq_loc_count = df.drop_duplicates(subset=['uid','lat','lng'])
uniq_loc_count =uniq_loc_count.groupby('uid').size().reset_index(name='uniq_loc_count')
print('No. of unique location at individuals\' activity clusters')
uniq_loc_count['uniq_loc_count'].quantile([.0, 0.25, .5, 0.75, 1])

cluster_count = df.groupby('uid')['cluster'].nunique().reset_index(name='cluster_count')
print('No. of unique activity location at individuals\' activity clusters')
cluster_count['cluster_count'].quantile([.0,0.25, .5,0.75,1])
