==> src/6.1-Generating-multiple-days.ipynb <==
%load_ext autoreload
%autoreload 2


%matplotlib inline

%cd C:\MAD4AG


for day in range(5):
    import os, datetime

    filename = os.path.join(os.getcwd(), r'dbs\twins\%s-simulation-%s-day.pkl'% (datetime.datetime.now().strftime('%m-%d'), day))

    %store filename
    %run ".\src\5.2-Matching-with-survey-twins_work_part.ipynb"
    %reset -f
